+++ ios/MullvadVPN/View controllers/Preferences/PreferencesDataSource.swift
@@ -375,7 +375,7 @@ final class PreferencesDataSource: UITableViewDiffableDataSource<
         case .wireGuardPorts:
             guard let view = tableView
                 .dequeueReusableHeaderFooterView(
-                    withIdentifier: HeaderFooterReuseIdentifiers.contentBlockerHeader.rawValue
+                    withIdentifier: HeaderFooterReuseIdentifiers.wireGuardPortHeader.rawValue
                 ) as? SettingsHeaderView else { return nil }
             configureWireguardPortsHeader(view)
             return view